+++ Sources/Sentry/SentryScreenshot.m
@@ -1,4 +1,5 @@
 #import "SentryScreenshot.h"
+#import "SentryCompiler.h"
 #import "SentryDependencyContainer.h"
 #import "SentryDispatchQueueWrapper.h"
 #import "SentryUIApplication.h"
@@ -43,11 +44,20 @@
     NSMutableArray *result = [NSMutableArray arrayWithCapacity:windows.count];
 
     for (UIWindow *window in windows) {
-        UIGraphicsBeginImageContext(window.frame.size);
+        CGSize size = window.frame.size;
+        if (size.width == 0 || size.height == 0) {
+            // avoid API errors reported as e.g.:
+            // [Graphics] Invalid size provided to UIGraphicsBeginImageContext(): size={0, 0},
+            // scale=1.000000
+            continue;
+        }
+        UIGraphicsBeginImageContext(size);
 
         if ([window drawViewHierarchyInRect:window.bounds afterScreenUpdates:false]) {
             UIImage *img = UIGraphicsGetImageFromCurrentImageContext();
-            if (img.size.width > 0 || img.size.height > 0) {
+            // this shouldn't happen now that we discard windows with either 0 height or 0 width,
+            // but still, we shouldn't send any images with either one.
+            if (LIKELY(img.size.width > 0 && img.size.height > 0)) {
                 NSData *bytes = UIImagePNGRepresentation(img);
                 if (bytes && bytes.length > 0) {
                     [result addObject:bytes];