+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 - Crash when serializing invalid objects (#2858)
+- Don't send screenshots with either width or height of 0 (#2876))
 
 ## 8.4.0